+++ opteryx/operators/show_functions_node.py
@@ -1,0 +1,60 @@
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Show Functions Node
+
+This is a SQL Query Execution Plan Node.
+"""
+from typing import Iterable
+import pyarrow
+
+from opteryx import operators, functions
+from opteryx.models import Columns, QueryDirectives, QueryStatistics
+from opteryx.operators import BasePlanNode
+
+
+class ShowFunctionsNode(BasePlanNode):
+    def __init__(
+        self, directives: QueryDirectives, statistics: QueryStatistics, **config
+    ):
+        super().__init__(directives=directives, statistics=statistics)
+        self._full = config.get("full")
+        self._extended = config.get("extended")
+
+    @property
+    def name(self):  # pragma: no cover
+        return "Show Functions"
+
+    @property
+    def config(self):  # pragma: no cover
+        return ""
+
+    def execute(self) -> Iterable:
+
+        buffer = []
+
+        for function in functions.functions():
+            buffer.append({"name": function, "type": "function"})
+        for aggregate in operators.aggregators():
+            buffer.append({"name": aggregate, "type": "aggregator"})
+
+        table = pyarrow.Table.from_pylist(buffer)
+        table = Columns.create_table_metadata(
+            table=table,
+            expected_rows=len(buffer),
+            name="show_columns",
+            table_aliases=[],
+        )
+
+        yield table
+        return